backend-server/requirements.txt
flask==2.3.3
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
flask-cors==4.0.0
pymongo==4.5.0
flask-pymongo==2.3.0
pytest==7.4.2
